backend/pyproject.toml
[tool.poetry]
name = "live-ai-assistant-backend"
version = "0.1.0"
description = "FastAPI backend for Live AI Assistant"
authors = ["Your Name <you@example.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
SQLAlchemy = { version = "^2.0.30", extras = ["asyncio"] }
asyncpg = "^0.29.0"
alembic = "^1.13.0"
redis = "^5.0.0"
httpx = "^0.27.0"
structlog = "^24.1.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
ruff = "^0.4.0"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

